+++ lib/generators/cms/attribute/attribute_generator.rb
@@ -9,7 +9,7 @@ module Cms
         type: :string,
         aliases: '-t',
         default: 'string',
-        desc: 'Type of the CMS attribute (string | text | html | markdown | enum | boolean | multienum | linklist | date).'
+        desc: 'Type of the CMS attribute (string | text | html | markdown | enum | boolean | multienum | linklist | date | float).'
 
       class_option :values,
         type: :array,